Parsing moves. Compiles but does not work yet.

+++ src-scala/src/main/scala/org/codingteam/icfpc2017/GameMap.scala
@@ -12,7 +12,7 @@ import scala.io.Source
 
 object GameMap {
 
-  type SiteId = Int
+  type SiteId = BigInt
 
   case class Site(id: SiteId)
